fitur-riwayat dan search

+++ routes/api.php
@@ -29,8 +29,8 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-// ->middleware('auth:sanctum','ablity:access-user');
 
+//auth
 Route::post('registerUser',[authController::class, 'RegisterUser']);//register
 Route::post('login',[authController::class, 'Login']);//login
 Route::post('logout',[authController::class, 'Logout'])->middleware('auth:sanctum');//logout
@@ -44,15 +44,16 @@ Route::post('ruangan/{id}',[RuanganController::class,'update'])->middleware('aut
 Route::get('peminjaman/submitted',[PeminjamanController::class, 'index'])->middleware('auth:sanctum','ablity:access-admin');//menampilkan data peminjaman yang belum disetujui
 Route::get('peminjaman/approve',[PeminjamanController::class, 'peminjamApprove'])->middleware('auth:sanctum','ablity:access-admin');//menampilkan data peminjaman approve
 Route::get('peminjaman/inprogress',[PeminjamanController::class, 'peminjamInProgress'])->middleware('auth:sanctum','ablity:access-admin');//menampilkan data peminjaman in progress
-Route::get('peminjaman/{id}',[PeminjamanController::class,'show'])->middleware('auth:sanctum');//menampilkan data berdasarkan id
+Route::get('peminjaman/riwayat',[PeminjamanController::class, 'riwayat'])->middleware('auth:sanctum','ablity:access-admin');//riwayat
+Route::get('peminjaman/{id}',[PeminjamanController::class,'show'])->middleware('auth:sanctum');//menampilkan data peminjaman berdasarkan id
 Route::get('unduhFileDokumen/{id}',[PeminjamanController::class,'unduhFile']);//
 Route::post('peminjaman',[PeminjamanController::class, 'store']);//tambah peminjaman
 Route::put('peminjaman/{id}',[PeminjamanController::class,'updateStatus'])->middleware('auth:sanctum','ablity:access-admin');//edit Status
-Route::delete('peminjaman/{id}',[PeminjamanController::class, 'destroy']);
+Route::delete('peminjaman/{id}',[PeminjamanController::class, 'destroy'])->middleware('auth:sanctum','ablity:access-user');//hapus peminjaman
 
 
 //user
 Route::get('peminjamanbyuser/{id}',[PeminjamanController::class,'peminjamanByUser'])->middleware('auth:sanctum','ablity:access-user');//menampilkan data berdasarkan id
 Route::post('EditPeminjaman/{id}',[PeminjamanController::class,'update'])->middleware('auth:sanctum','ablity:access-user');//edit ruangan
-Route::patch('peminjaman/{id}/feedback',[PeminjamanController::class,'feedback'])->middleware('auth:sanctum','ablity:access-user');
-
+Route::patch('peminjaman/{id}/feedback',[PeminjamanController::class,'feedback'])->middleware('auth:sanctum','ablity:access-user');//
+Route::get('peminjaman/riwayat/{id}',[PeminjamanController::class, 'riwatyatPeminjamanByUser'])->middleware('auth:sanctum','ablity:access-user');//riwayat peminjaman user
